+++ typescript/infra/config/docker.ts
@@ -3,9 +3,7 @@ const GHCR_REGISTRY = 'ghcr.io/hyperlane-xyz';
 export const DockerImageNames = {
   AGENT: 'hyperlane-agent',
   MONOREPO: 'hyperlane-monorepo',
-  KEY_FUNDER: 'hyperlane-key-funder',
-  WARP_MONITOR: 'hyperlane-warp-monitor',
-  REBALANCER: 'hyperlane-rebalancer',
+  NODE_SERVICES: 'hyperlane-node-services',
 } as const;
 
 type DockerImageReposType = {
@@ -39,17 +37,17 @@ interface MainnetDockerTags extends BaseDockerTags {
 
 export const mainnetDockerTags: MainnetDockerTags = {
   // rust agents
-  relayer: '6d8f392-20260310-131412',
-  relayerRC: '6d8f392-20260310-131412',
+  relayer: '1d56bfd-20260312-161015',
+  relayerRC: '1d56bfd-20260312-161015',
   validator: '6d8f392-20260310-131412',
   validatorRC: '6d8f392-20260310-131412',
   scraper: '6d8f392-20260310-131412',
   // monorepo services
   checkWarpDeploy: 'main',
   // standalone services
-  keyFunder: 'c558a9f-20260304-105251',
-  warpMonitor: 'c558a9f-20260304-105251',
-  rebalancer: 'c558a9f-20260304-105251',
+  keyFunder: '3b17358-20260315-183126',
+  warpMonitor: '3b17358-20260315-183126',
+  rebalancer: '3b17358-20260315-183126',
 };
 
 export const testnetDockerTags: BaseDockerTags = {
@@ -60,5 +58,5 @@ export const testnetDockerTags: BaseDockerTags = {
   validatorRC: 'c558a9f-20260304-105241',
   scraper: 'c558a9f-20260304-105241',
   // standalone services
-  keyFunder: 'c558a9f-20260304-105251',
+  keyFunder: '3b17358-20260315-183126',
 };