+++ typescript/sdk/src/index.ts
@@ -706,9 +706,12 @@ export { DeployableTokenType, TokenType } from './token/config.js';
 export {
   expandVirtualWarpDeployConfig,
   expandWarpDeployConfig,
+  filterWarpCoreConfigMapByChains,
+  getChainsFromWarpCoreConfig,
   getRouterAddressesFromWarpCoreConfig,
   splitWarpCoreAndExtendedConfigs,
   transformConfigToCheck,
+  warpCoreConfigMatchesChains,
 } from './token/configUtils.js';
 export {
   hypERC20contracts,
@@ -797,9 +800,12 @@ export {
   EverclearCollateralTokenConfig,
   EverclearEthBridgeTokenConfig,
   isXERC20TokenConfig,
-  isMultiCollateralTokenConfig,
-  MultiCollateralTokenConfig,
+  isCrossCollateralTokenConfig,
+  isOftTokenConfig,
+  CrossCollateralTokenConfig,
   NativeTokenConfig,
+  OftTokenConfig,
+  OftTokenConfigSchema,
   NativeTokenConfigSchema,
   SyntheticRebaseTokenConfig,
   SyntheticRebaseTokenConfigSchema,