+++ typescript/cosmos-types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/cosmos-types
 
+## 27.1.0
+
 ## 27.0.0
 
 ## 26.0.0