+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @hyperlane-xyz/infra
 
+## 27.1.0
+
+### Patch Changes
+
+- Updated dependencies [de5f6b5]
+- Updated dependencies [a1f9e41]
+- Updated dependencies [b892e61]
+- Updated dependencies [b892e61]
+- Updated dependencies [abdbbf5]
+- Updated dependencies [b892e61]
+- Updated dependencies [7af7728]
+  - @hyperlane-xyz/sdk@27.1.0
+  - @hyperlane-xyz/provider-sdk@3.1.0
+  - @hyperlane-xyz/deploy-sdk@3.1.0
+  - @hyperlane-xyz/utils@27.1.0
+  - @hyperlane-xyz/http-registry-server@27.1.0
+  - @hyperlane-xyz/keyfunder@0.1.9
+  - @hyperlane-xyz/metrics@0.2.2
+  - @hyperlane-xyz/rebalancer@27.1.0
+  - @hyperlane-xyz/relayer@1.1.9
+  - @hyperlane-xyz/tron-sdk@22.1.1
+  - @hyperlane-xyz/core@11.0.2
+
 ## 27.0.0
 
 ### Patch Changes